backend/requirements.txt
# G Telepathy — Backend Dependencies

# Web Framework
fastapi==0.111.0
uvicorn[standard]==0.29.0
python-multipart==0.0.9

# Real-time / WebSockets
python-socketio==5.11.2
websockets==12.0

# Database / Auth
supabase==2.4.3
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4

# Google Cloud APIs
google-cloud-translate==3.15.3
google-cloud-speech==2.26.0

# Voice Cloning
elevenlabs==1.2.1

# Encryption
pynacl==1.5.0

# HTTP Client
httpx==0.27.0
aiohttp==3.9.5

# Environment
python-dotenv==1.0.1
pydantic-settings==2.2.1

# Utilities
python-dateutil==2.9.0
